+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/State/Aura_AutoBlue_Short_State_Koach.java
@@ -65,6 +65,7 @@ import com.acmerobotics.roadrunner.TranslationalVelConstraint;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.hardware.VoltageSensor;
@@ -105,6 +106,7 @@ import java.util.concurrent.TimeUnit;
 
 @Config
 @Autonomous(name="Blue_Short_State_Koach", group="Linear OpMode")
+@Disabled
 
 public class Aura_AutoBlue_Short_State_Koach extends LinearOpMode {
