+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/State/Aura_AutoRed_Long_State_No.java
@@ -55,6 +55,7 @@ import com.acmerobotics.roadrunner.SequentialAction;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -93,6 +94,7 @@ import java.util.concurrent.TimeUnit;
 
 @Config
 @Autonomous(name="Red_Long_State_No", group="Linear OpMode")
+@Disabled
 
 public class Aura_AutoRed_Long_State_No extends LinearOpMode {
