+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/State/Aura_AutoRed_Short_State_No.java
@@ -91,9 +91,9 @@ import java.util.concurrent.TimeUnit;
  * Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
  */
 
-@Disabled
 @Config
 @Autonomous(name="Red_Short_State_No", group="Linear OpMode")
+@Disabled
 
 public class Aura_AutoRed_Short_State_No extends LinearOpMode {
